Improved ranking template looks
