+++ src/Gateways/Desktop/Core/Components/Styles/Border.xaml
@@ -41,4 +41,33 @@
         <Setter Property="Height" Value="40px" />
     </Style>
 
+    <Style x:Key="Cell" TargetType="Border">
+        <Setter Property="Padding" Value="5px 0" />
+        <Setter Property="BorderThickness" Value="1,1,1,1" />
+        <Setter Property="BorderBrush" Value="Black" />
+    </Style>
+
+    <Style x:Key="Cell-Thick-Right" TargetType="Border" BasedOn="{StaticResource Cell}">
+        <Setter Property="BorderThickness" Value="0,0,0,1" />
+    </Style>
+
+    <Style x:Key="Cell-No-Padding" TargetType="Border" BasedOn="{StaticResource Cell}">
+        <Setter Property="Padding" Value="0" />
+    </Style>
+
+    <Style x:Key="Border-Heading" BasedOn="{StaticResource Cell}" TargetType="Border">
+        <Setter Property="Background" Value="DimGray" />
+        <Setter Property="Height" Value="50px" />
+    </Style>
+
+    <Style  x:Key="Border-Left" BasedOn="{StaticResource Border-Horizontal}" TargetType="Border">
+        <Setter Property="HorizontalAlignment" Value="Left" />
+        <Setter Property="Padding" Value="50px 0" />
+    </Style>
+
+    <Style  x:Key="Border-Right" BasedOn="{StaticResource Border-Horizontal}" TargetType="Border">
+        <Setter Property="HorizontalAlignment" Value="Right" />
+        <Setter Property="Padding" Value="50px 0" />
+    </Style>
+
 </ResourceDictionary>
