+++ src/Gateways/Desktop/Core/Components/Styles/ListBox.xaml
@@ -1,11 +1,19 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
-    <Style x:Key="ListBox-Vertical" TargetType="ListBox">
+    <Style x:Key="ListBox" TargetType="ListBox">
+        <Setter Property="BorderThickness" Value="0" />
+    </Style>
+
+    <Style x:Key="ListBox-Vertical" TargetType="ListBox" BasedOn="{StaticResource ListBox}">
         <Setter Property="HorizontalContentAlignment" Value="Stretch" />
     </Style>
 
-    <Style x:Key="ListBox-Horizontal" TargetType="ListBox">
+    <Style x:Key="ListBox-Vertical-Border" TargetType="ListBox" BasedOn="{StaticResource ListBox-Vertical}">
+        <Setter Property="BorderThickness" Value="1" />
+    </Style>
+
+    <Style x:Key="ListBox-Horizontal" TargetType="ListBox" BasedOn="{StaticResource ListBox}">
         <Setter Property="VerticalContentAlignment" Value="Stretch" />
         <Setter Property="ItemsPanel">
             <Setter.Value>
@@ -14,6 +22,10 @@
                 </ItemsPanelTemplate>
             </Setter.Value>
         </Setter>
+    </Style>
+
+    <Style x:Key="ListBox-Horizontal-Border" TargetType="ListBox" BasedOn="{StaticResource ListBox-Horizontal}">
+        <Setter Property="BorderThickness" Value="1" />
     </Style>
 
 </ResourceDictionary>
