+++ src/Gateways/Desktop/Views/Content/Common/Participants/ParticipantTemplateModel.cs
@@ -1,28 +1,22 @@
-﻿using EnduranceJudge.Core.Mappings;
-using EnduranceJudge.Core.Utilities;
+﻿using EnduranceJudge.Core.Utilities;
 using EnduranceJudge.Domain.State.Participants;
-using EnduranceJudge.Domain.State.Participations;
 using EnduranceJudge.Gateways.Desktop.Core;
 using EnduranceJudge.Gateways.Desktop.Services;
 using EnduranceJudge.Gateways.Desktop.Views.Content.Common.PhasePerformances;
 using EnduranceJudge.Localization.Translations;
 using Prism.Commands;
-using System.Collections.ObjectModel;
-using System.Linq;
 using System.Windows;
 using System.Windows.Media;
 
 namespace EnduranceJudge.Gateways.Desktop.Views.Content.Common.Participants;
 
-public class ParticipantTemplateModel : ViewModelBase, ICollapsable
+public class ParticipantTemplateModel : ParticipantTemplateModelBase<PerformanceTemplateModel>, ICollapsable
 {
     private readonly IPrinter printer;
 
-    public ParticipantTemplateModel(Participant participant, bool isExpanded = true)
+    public ParticipantTemplateModel(Participant participant, bool isExpanded = true) : base(participant)
     {
         this.printer = StaticProvider.GetService<IPrinter>();
-        this.Number = participant.Number;
-        this.UpdatePhases(participant.Participation);
         this.ToggleVisibility = new DelegateCommand(this.ToggleVisibilityAction);
         this.Print = new DelegateCommand<Visual>(this.PrintAction);
         if (isExpanded)
@@ -36,28 +30,12 @@ public class ParticipantTemplateModel : ViewModelBase, ICollapsable
     private string toggleText = Words.EXPAND;
     private readonly int number;
     private Visibility visibility = Visibility.Collapsed;
-    public ObservableCollection<PerformanceTemplateModel> Performances { get; } = new();
-    public ObservableCollection<double> PhaseLengths { get; } = new();
-
-    private void UpdatePhases(Participation participation)
-    {
-        var lengths = participation.Performances.Select(x => x.Phase.LengthInKm);
-        var viewModels = participation.Performances.MapEnumerable<PerformanceTemplateModel>();
-
-        this.Performances.AddRange(viewModels);
-        this.PhaseLengths.AddRange(lengths);
-    }
 
     private void PrintAction(Visual control)
     {
         this.printer.Print(control);
     }
 
-    public int Number
-    {
-        get => this.number;
-        private init => this.SetProperty(ref this.number, value);
-    }
     public Visibility Visibility
     {
         get => this.visibility;
