+++ src/Gateways/Desktop/Views/Content/Configuration/Children/Competitions/CompetitionView.xaml
@@ -92,7 +92,7 @@
 
             <Border Style="{StaticResource Border-Vertical}"
                     Grid.Row="5" Grid.Column="1">
-                <ListBox Style="{StaticResource ListBox-Vertical}"
+                <ListBox Style="{StaticResource ListBox-Vertical-Border}"
                          ItemsSource="{Binding Phases}" />
             </Border>
 
@@ -112,7 +112,7 @@
             <Border Style="{StaticResource Border-Vertical}"
                     Grid.Row="7" Grid.Column="1" >
                 <DockPanel Style="{StaticResource Stretch-Last}">
-                    <ListBox Style="{StaticResource ListBox-Vertical}"
+                    <ListBox Style="{StaticResource ListBox-Vertical-Border}"
                              ItemsSource="{Binding Participants}"
                              Name="Personnel"/>
                 </DockPanel>
