+++ src/Gateways/Desktop/Views/Content/Configuration/Roots/EnduranceEvents/EnduranceEventView.xaml
@@ -106,7 +106,7 @@
             <Border Style="{StaticResource Border-Vertical}"
                     Grid.Row="8" Grid.Column="1">
                 <DockPanel Style="{StaticResource Stretch-Last}">
-                    <ListBox Style="{StaticResource ListBox-Vertical}"
+                    <ListBox Style="{StaticResource ListBox-Vertical-Border}"
                              ItemsSource="{Binding Competitions}" />
                 </DockPanel>
             </Border>
