+++ src/Gateways/Desktop/Views/Content/Configuration/Roots/Horses/Listing/HorseListView.xaml
@@ -47,7 +47,7 @@
                 Grid.Row="3">
             <ScrollViewer Style="{StaticResource Scroll}"
                           PreviewMouseWheel="HandleScroll">
-                <ListBox Style="{StaticResource ListBox-Vertical}"
+                <ListBox Style="{StaticResource ListBox-Vertical-Border}"
                          ItemsSource="{Binding ListItems}" />
             </ScrollViewer>
         </Border>
