+++ src/Gateways/Desktop/Views/Content/Rankings/PerformanceResults/PerfomanceResultTemplate.xaml
@@ -1,0 +1,67 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:performanceResults="clr-namespace:EnduranceJudge.Gateways.Desktop.Views.Content.Rankings.PerformanceResults"
+                    xmlns:localization="clr-namespace:EnduranceJudge.Localization.Translations;assembly=EnduranceJudge.Localization">
+
+    <DataTemplate DataType="{x:Type performanceResults:PerformanceResultTemplateModel}">
+        <StackPanel Style="{StaticResource Performance-Column}" Margin="-5px -2.5px">
+            <!-- Arrival Time-->
+            <Border Style="{StaticResource Cell}"
+                    Grid.Row="1" Grid.Column="1">
+                <TextBlock Style="{StaticResource Text}"
+                           Text="{Binding ArrivalTimeString}" />
+            </Border>
+            <!-- Inspection Time -->
+            <Border Style="{StaticResource Cell}"
+                    Grid.Row="2" Grid.Column="1">
+                <TextBlock Style="{StaticResource Text}"
+                           Text="{Binding InspectionTimeString}" />
+            </Border>
+            <!-- Re-inspection Time -->
+            <Border Style="{StaticResource Cell}"
+                    Grid.Row="3" Grid.Column="1">
+                <TextBlock Style="{StaticResource Text}"
+                           Text="{Binding ReInspectionTimeString}" />
+            </Border>
+            <!-- Required inspection Time -->
+            <Border Style="{StaticResource Cell}"
+                    Grid.Row="2" Grid.Column="1">
+                <TextBlock Style="{StaticResource Text}"
+                           Text="{Binding RequiredInspectionTimeString}" />
+            </Border>
+            <!-- Loop span -->
+            <Border Style="{StaticResource Cell}"
+                    Grid.Row="3" Grid.Column="1">
+                <TextBlock Style="{StaticResource Text}">
+                    <Run Text="{Binding LoopSpan, Mode=OneWay, StringFormat=hh\\:mm}" />
+                    <Run Text="{x:Static localization:Words.HOURS_SYMBOL}" />
+                </TextBlock>
+            </Border>
+            <!-- Phase Span -->
+            <Border Style="{StaticResource Cell}"
+                    Grid.Row="4" Grid.Column="1">
+                <TextBlock Style="{StaticResource Text}">
+                    <Run Text="{Binding PhaseSpan, Mode=OneWay, StringFormat=hh\\:mm}" />
+                    <Run Text="{x:Static localization:Words.HOURS_SYMBOL}" />
+                </TextBlock>
+            </Border>
+            <!-- Average Loop Speed -->
+            <Border Style="{StaticResource Cell}"
+                    Grid.Row="5" Grid.Column="1">
+                <TextBlock Style="{StaticResource Text}">
+                    <Run Text="{Binding AverageSpeedForLoopInKpH, Mode=OneWay, StringFormat=N2}" />
+                    <Run Text="{x:Static localization:Words.KM_PER_HOUR}" />
+                </TextBlock>
+            </Border>
+            <!-- Average Phase Speed -->
+            <Border Style="{StaticResource Cell}"
+                    Grid.Row="6" Grid.Column="1">
+                <TextBlock Style="{StaticResource Text}">
+                    <Run Text="{Binding AverageSpeedForPhaseInKpH, Mode=OneWay, StringFormat=N2}" />
+                    <Run Text="{x:Static localization:Words.KM_PER_HOUR}" />
+                </TextBlock>
+            </Border>
+        </StackPanel>
+    </DataTemplate>
+
+</ResourceDictionary>
