+++ src/Gateways/Desktop/Views/Content/Rankings/RankLists/RankListTemplate.xaml
@@ -3,61 +3,71 @@
                     xmlns:local="clr-namespace:EnduranceJudge.Gateways.Desktop.Views.Content.Rankings.RankLists">
 
     <DataTemplate DataType="{x:Type local:RankListTemplateModel}">
-        <Grid Style="{StaticResource Grid}">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="80px" />
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="Auto" />
-            </Grid.RowDefinitions>
-
-            <!-- Header -->
-            <Border Style="{StaticResource Border-Vertical}"
-                    Grid.Row="0">
-                <StackPanel Style="{StaticResource List-Horizontal}">
-
-                    <!-- Event  -->
-                    <StackPanel Style="{StaticResource List-Vertical}">
-
-                        <TextBlock Style="{StaticResource Text-Bold}"
-                                   Text="{Binding EventName, Mode=OneWay}" />
-
-                        <TextBlock Style="{StaticResource Text-Bold}">
-                            <Run Style="{StaticResource Run-Bold}"
-                                 Text="{Binding PopulatedPlace, Mode=OneWay}" />
-                            <Run Style="{StaticResource Run-Bold}"
-                                 Text=" - " />
-                            <Run Style="{StaticResource Run-Bold}"
-                                 Text="{Binding CountryName, Mode=OneWay}" />
-                        </TextBlock>
-
-                        <TextBlock Style="{StaticResource Text-Bold}"
-                                   Text="{Binding CompetitionDate, Mode=OneWay}" />
-                    </StackPanel>
-
-                    <!-- Organizers -->
-                    <TextBlock Style="{StaticResource Text-H3}"
-                               Text="{Binding Organizer, Mode=OneWay}"/>
-
-                </StackPanel>
-            </Border>
-
-            <!-- Title -->
-            <Border Style="{StaticResource Border-Vertical}"
-                    Grid.Row="1"
-                    BorderThickness="1"
-                    BorderBrush="Black"
-                    Background="DimGray">
-                <TextBlock Style="{StaticResource Text-H2}"
-                           Text="{Binding Title, Mode=OneWay}" />
-            </Border>
-
-            <!-- Rank List -->
-            <Border Style="{StaticResource Border-Vertical}"
-                    Grid.Row="2">
-                <ListBox Style="{StaticResource ListBox-Vertical}"
-                         ItemsSource="{Binding RankList, Mode=OneWay}" />
-            </Border>
-        </Grid>
+        <Border Style="{StaticResource Cell}" Padding="0">
+            <Grid Style="{StaticResource Grid}">
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="80px" />
+                    <RowDefinition Height="Auto" />
+                    <RowDefinition Height="Auto" />
+                    <RowDefinition Height="Auto" />
+                    <RowDefinition Height="Auto" />
+                </Grid.RowDefinitions>
+                <!-- Header -->
+                <Border Style="{StaticResource Border-Vertical}"
+                        Grid.Row="0">
+                    <DockPanel Style="{StaticResource Dock-Horizontal}">
+                        <!-- Event  -->
+                        <StackPanel Style="{StaticResource List-Vertical}">
+                            <TextBlock Style="{StaticResource Text-Bold}"
+                                       Text="{Binding EventName, Mode=OneWay}" />
+                            <TextBlock Style="{StaticResource Text-Bold}">
+                                <Run Style="{StaticResource Run-Bold}"
+                                     Text="{Binding PopulatedPlace, Mode=OneWay}" />
+                                <Run Style="{StaticResource Run-Bold}"
+                                     Text=" - " />
+                                <Run Style="{StaticResource Run-Bold}"
+                                     Text="{Binding CountryName, Mode=OneWay}" />
+                            </TextBlock>
+                            <TextBlock Style="{StaticResource Text-Bold}"
+                                       Text="{Binding CompetitionDate, Mode=OneWay}" />
+                        </StackPanel>
+                        <!-- Organizers -->
+                        <TextBlock Style="{StaticResource Text-H3-Right}"
+                                   Text="{Binding Organizer, Mode=OneWay}"/>
+                    </DockPanel>
+                </Border>
+                <!-- Title -->
+                <Border Style="{StaticResource Border-Heading}"
+                        Grid.Row="1">
+                    <TextBlock Style="{StaticResource Text-H2}" TextAlignment="Center"
+                               Text="{Binding Title, Mode=OneWay}" />
+                </Border>
+                <!-- Rank List -->
+                <Border Style="{StaticResource Border-Vertical}"
+                        Grid.Row="2">
+                    <ListBox Style="{StaticResource ListBox-Vertical}"
+                             ItemsSource="{Binding RankList, Mode=OneWay}" />
+                </Border>
+                <!-- Footer -->
+                <Border Style="{StaticResource Border-Heading}"
+                        Grid.Row="3">
+                    <DockPanel Style="{StaticResource Dock-Horizontal}">
+                        <!-- President Ground Jury -->
+                        <TextBlock Style="{StaticResource Text-Align-Left}"
+                                   Text="{Binding PresidentGroundJuryName}" />
+                        <!-- Chief Steward -->
+                        <TextBlock Style="{StaticResource Text-Align-Right}"
+                                   Text="{Binding ChiefStewardName}" />
+                    </DockPanel>
+                </Border>
+                <!-- Date -->
+                <Border Style="{StaticResource Border-Vertical}"
+                        Grid.Row="4">
+                    <TextBlock Style="{StaticResource Text-Align-Right}"
+                               Text="{Binding DateNowString}" />
+                </Border>
+            </Grid>
+        </Border>
     </DataTemplate>
 
 </ResourceDictionary>
