+++ src/Gateways/Desktop/Views/Templates.xaml
@@ -12,7 +12,8 @@
         <ResourceDictionary Source="Content/Common/PhasePerformances/PerformanceTemplate.xaml" />
         <ResourceDictionary Source="Content/Common/Participants/ParticipantTemplate.xaml" />
         <ResourceDictionary Source="Content/Rankings/RankLists/RankListTemplate.xaml" />
-        <ResourceDictionary Source="Content/Rankings/RankListEntries/RankListEntryTemplate.xaml" />
+        <ResourceDictionary Source="Content/Rankings/ParticipantResults/ParticipantResultTemplate.xaml" />
+        <ResourceDictionary Source="Content/Rankings/PerformanceResults/PerfomanceResultTemplate.xaml" />
     </ResourceDictionary.MergedDictionaries>
 
 </ResourceDictionary>
